Replace node.set with node.default in server recipe
The primary difference being that node.set is persisted with the node object,
whereas node.default is built during each chef run. Since this usage is using
`search_for_nodes` to discover the value during the chef run node.default seems
like it should be safe.

This is in line with current community recommendations and partially fixes #80