Fixing service file check to support systemd (#82)
* Change service check for additional OSes

Ran into an issue in production, using the sidecar recipe on Ubuntu 16.04, it failed if the sidecar was already installed because the file check didn't exist for systemd. The same goes for [RHEL|OEL|CENTOS]7. There may be a better way to do this.

* Fixing spacing issues

Fix build issues with spaces....

* Proper syntax changes

file.exists? reports deprecated from rubocop, so replacing it with file.exist?