Adding extra network example config in README (#97)
I've found that this piece of config was very useful, it might save a good few hours for users deploying to remote machines behind a NAT. Please add more detail if you can, I have mainly figured this out by trial and error. I know there is an alternative way to set it up by using the same port, but it didn't work in my case.