Allow skip java version check
This cookbook should obviously fail if java version provided is not compliant
(<8, >11) but one should be able to get through this security, in case of a uncommon java install.
Ex:
Ohai retrieves the version of oracle java 8 to 1.8 (which is correct), but the check used in the cookbook
does not allow it
Ohai is the worst, it sets its attributes to the highest precedence possible, cf https://docs.chef.io/attribute_precedence/
So, event an node.automatic['languages']['java']['version'] = '11.0.15' before calling the `include_recipe 'graylog2'`,
could not do the trick.

A simple solution is to allow the cookbook user to by pass this security with:
```
node.override['graylog2']['server']['skip_java_version_check'] = true
```

Obviously, this setting is set to false by default.